+++ packages/calcite-components/src/components/inline-editable/inline-editable.e2e.ts
@@ -1,7 +1,17 @@
 // @ts-strict-ignore
 import { E2EPage, newE2EPage } from "@arcgis/lumina-compiler/puppeteerTesting";
 import { beforeEach, describe, expect, it } from "vitest";
-import { accessible, defaults, disabled, hidden, labelable, renders, t9n, themed } from "../../tests/commonTests";
+import {
+  accessible,
+  defaults,
+  focusable,
+  disabled,
+  hidden,
+  labelable,
+  renders,
+  t9n,
+  themed,
+} from "../../tests/commonTests";
 import { html } from "../../../support/formatting";
 import type { Input } from "../input/input";
 import { findAll } from "../../tests/utils/puppeteer";
@@ -42,6 +52,19 @@ describe("calcite-inline-editable", () => {
         </calcite-inline-editable>
       `,
       { focusTarget: { tab: "calcite-inline-editable", click: "calcite-input" } },
+    );
+  });
+
+  describe("focusable", () => {
+    focusable(
+      html`
+        <calcite-inline-editable>
+          <calcite-input />
+        </calcite-inline-editable>
+      `,
+      {
+        focusTargetSelector: "calcite-input",
+      },
     );
   });