+++ packages/calcite-components/src/components/navigation-logo/navigation-logo.tsx
@@ -1,8 +1,8 @@
 // @ts-strict-ignore
 import { h, Fragment, JsxNode, LitElement, method, property } from "@arcgis/lumina";
-import { componentFocusable } from "../../utils/component";
 import { Heading, HeadingLevel } from "../functional/Heading";
 import { IconNameOrString } from "../icon/interfaces";
+import { useSetFocus } from "../../controllers/useSetFocus";
 import { CSS } from "./resources";
 import { styles } from "./navigation-logo.scss";
 
@@ -18,6 +18,12 @@ export class NavigationLogo extends LitElement {
   static override shadowRootOptions = { mode: "open" as const, delegatesFocus: true };
 
   static override styles = styles;
+
+  // #endregion
+
+  // #region Private Properties
+
+  private focusSetter = useSetFocus<this>()(this);
 
   // #endregion
 
@@ -68,13 +74,20 @@ export class NavigationLogo extends LitElement {
 
   // #region Public Methods
 
-  /** Sets focus on the component. */
+  /**
+   * Sets focus on the component.
+   *
+   * @param options - When specified an optional object customizes the component's focusing process. When `preventScroll` is `true`, scrolling will not occur on the component.
+   *
+   * @mdn [focus(options)](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/focus#options)
+   */
   @method()
-  async setFocus(): Promise<void> {
-    await componentFocusable(this);
-    if (this.href) {
-      this.el.focus();
-    }
+  async setFocus(options?: FocusOptions): Promise<void> {
+    return this.focusSetter(() => {
+      if (this.href) {
+        return this.el;
+      }
+    }, options);
   }
 
   // #endregion